+++ cmdexec/js.wasm.go
@@ -12,6 +12,9 @@ func RunVMCode(jsCode string, vm *goja.Runtime) {
 	js.Global().Call("eval", jsCode)
 }
 
+//  The GenerateChainCode function creates a JavaScript code snippet
+//  that attempts to parse a given HTTP response body as JSON, logging
+//  success or failure, and stores the result accordingly.
 func GenerateChainCode(httpRespBody string) string {
 	code := `try {
 		result = JSON.parse(String.raw` + "`" + httpRespBody + "`" + `)
